+++ README.md
@@ -3,5 +3,5 @@ An application to showcase and execute Splashkit games
 ## Install
 + Install the [SplashKit](https://splashkit.io) SDK using the [guide](https://splashkit.io/articles/installation/)
 ## Run
-+ Compile the application with the command ```skm clang++ program.cpp -o test -lstdc++fs```  
++ Compile the application with the command ```skm clang++ program.cpp -o test -Iinclude -lstdc++fs```  
 + Run the application ```./test```
